// src/shared/api/generated/models/endUserCaseCommandErrorDetailsResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { EndUserCaseCommandErrorDetailsResponseDtoAvailableStatusesItem } from "./endUserCaseCommandErrorDetailsResponseDtoAvailableStatusesItem";
import type { EndUserCaseCommandCurrentResponseDto } from "./endUserCaseCommandCurrentResponseDto";
import type { EndUserCaseCommandErrorDetailsResponseDtoCurrentStatus } from "./endUserCaseCommandErrorDetailsResponseDtoCurrentStatus";
import type { EndUserCaseCommandErrorDetailsResponseDtoEscalationStatus } from "./endUserCaseCommandErrorDetailsResponseDtoEscalationStatus";
import type { EndUserCaseCommandValidationIssueResponseDto } from "./endUserCaseCommandValidationIssueResponseDto";
import type { EndUserCasePriorityPolicyResponseDto } from "./endUserCasePriorityPolicyResponseDto";

export interface EndUserCaseCommandErrorDetailsResponseDto {
  availableStatuses?: EndUserCaseCommandErrorDetailsResponseDtoAvailableStatusesItem[];
  currentCase?: EndUserCaseCommandCurrentResponseDto;
  currentStatus?: EndUserCaseCommandErrorDetailsResponseDtoCurrentStatus;
  /** @minimum 1 */
  currentVersion?: number;
  escalationId?: string;
  escalationStatus?: EndUserCaseCommandErrorDetailsResponseDtoEscalationStatus;
  issues?: EndUserCaseCommandValidationIssueResponseDto[];
  mergedIntoCaseId?: string;
  priorityPolicy?: EndUserCasePriorityPolicyResponseDto;
}